fix: pushed main redirection funtion into utility file
